Tighen read_unauthenticated() view config
Explicitly state that the read_unauthenticated() view should only be
called when the user is not logged in.

Previously this view could have been called when the user _was_ logged
in, and the only reason that it wasn't was that read(), which is more
specific, would be called instead.

If the read() view were deleted then read_unauthenticated() would be
called even for authenticated requests.

I think it's good to 1. Not have to read multiple @view_configs to
determinse when one particular view will and will not be called, and
2. Be more explicit in @view_config and @view_defaults about when a view is and
is not intended to be called.