Move GroupSearchController to group_read route
Fix https://github.com/hypothesis/h/issues/4004

Change the URL of the group activity search page to
/groups/{pubid}/{slug}, the same as the other group pages.

This means that when the user is logged in, is a member of the group,
and has the search_page feature flag on, a request to
/groups/{pubid}/{slug} will go to
h.views.activity.GroupSearchController. It will then be routed to one of
GroupSearchController's methods depending on, for example, what params
are in the request.

If the feature flag is _off_ then the same request would go to
h.views.groups.read() (the classic group page) instead.

Even if the feature flag is on, if the user isn't a member of the group
or isn't logged in a request to /groups/{pubid}/{slug} will go to
h.views.groups.GroupJoinContoller or
h.views.groups.read_unauthenticated().

GroupSearchController is called _only_ if the user is logged in, a
member of the group, and has the search_page feature flag on.

Changes:

Change the route_name of GroupSearchController to "group_read", and add
view predicates so that it's only used when the "search_page" feature
flag is on and the user is a member of the group.

Also tighten up the view predicates of h.views.groups.read() (the
classic group read page) so that it's only called when the "search_page"
feature flag is off. This isn't necessary - Pyramid will call
GroupSearchController when the flag is on anyway - but I think it's
better to be more explicit and to be able to see, by looking at read()'s
own @view_config only, exactly when it will and won't be called.

Completely remove the "activity.group_search" route.

Replace all references to "activity.group_search" with "group_read".