Fix user and group search pages on iOS Safari
Fix https://github.com/hypothesis/h/issues/4116

When submitting the search form using the keyboard on the
/groups/{pubid}/{slug} page or the /users/{username} page on mobile
Safari, the group or user lozenge was being incorrectly deleted and the
browser incorrectly redirected to the /search page. On other browsers
this was not happening.

There is an `<input type="submit">` that's hidden with CSS that is the
submit button that's submitted when you hit Enter in the search form.
The problem is that mobile Safari was not finding this submit button,
skipping over it and instead finding the delete button on the user or
group lozenge and submitting that instead.

The reason turns out to be that display: none; was used in CSS to hide
the hidden input, and that causes Safari to ignore it for form
submission purposes. (This seems like a bug in mobile Safari - CSS
probably shouldn't affect form behavior.)

The fix is to use visibility: hidden; instead. That then breaks the
search box layout, so add position: absolute; to fix it again.