Make all search page forms use GET
Make all the forms and buttons on the search (and group and user search)
pages use GET rather than POST requests.

GET is more semantically correct for submitting a search form (because
the requests are idempotent / they don't change anything on the server). This
is also more consistent as some of the search form fields and buttons
were using GET and some POST - now they're all GET.

The one button on the search page that remains a POST request is the
"Leave group" button on the group search page - this is semantically
correct as a POST.

For most buttons this is simply a change from "GET" to "POST" and the
Python code now needs to look at request.params instead of request.POST.

The one exception is the "More info" button - this button no longer
requires a server-side redirect that it had, so that Python code is
removed.

This does result in one behavior change - when on the more info page, if
you submit the search form (by hitting enter in the search box or by
clicking on a tag or username) when the page reloads you will be back on
the search results page looking at the updated results. This is arguably
an improvement.