Set preferred/main width of annotation cards to 500px on desktop
Set the initial size (width) of the annotation card list to 500px on
desktop, rather than sizing based on the content of the card.

This fixes a problem where the card would be narrow if the bucket had a
long URL and only short quotes/bodies in the annotations.

On small screens the initial size (height) is still set based on the
content, using `auto` rather than `content` for compatibility with older
browsers.