Move nav bar margins onto content div
The nav bar's left and right margins were on its first and last child
elements, rather than on the containing element:

The first child element in the `<header class="nav-bar">` element
(the `<img class="nav-bar__logo">` had a `margin-left: 30px;` on it.
The last child element in the `<header class="nav-bar">`
(the `<div class="nav-bar-links">`) had a `margin-right: 60px;` on it.

Move both the `margin-left` and `margin-right` of these first and last
child elements onto their containing
`<div class="nav-bar__content">` element.

It seems tidier to have the outer margins on the container element,
and this is also consistent with how the matching margins on the
`<div class="search-result-container">` are done as well.

This is just a refactoring and shouldn't change how the page appears at all.