Fix right-alignment of search results
The `<div class="search-result-container">` (which contains the search
results and, on user and group pages, the search sidebar) had a
`margin-right: 30px;`. This is supposed to line up with the right margin
of the `<header class="nav-bar">` but it didn't - the header's right
margin is 60px.

So fix the search result container's margin to match the header's.