Center search page contents on screen
Fixes https://github.com/hypothesis/h/issues/4018

When the viewport is wider than the max width of the search page
contents then horizontally center the contents in the middle of the
viewport, instead of having some of the contents left-aligned and some
pushed all the way along to the right edge of the viewport (however far
away from the rest of the contents that might be).

Change some margins to padding, so that margin-left:auto and
margin-right:auto can be used to center things.

Add max-widths to the nav bar and search results containers, and add
margin-left:auto and margin-right:auto to center them when the viewport
is wider than this max-width.

The max-width is calculated as the max-width of the search results
container (950px), plus the search results container's margin-right (the
gap between the search results and the sidebar, 30px) plus the width of
the sidebar (285px) = 1265px.