Replace some hardcoded values with SCSS variables
We have a bunch of hardcoded values that are co-dependent: the max-width
of the nav bar and search bar containers need to be the same, and this
max width needs to be equal to the sum of the max width of the search
results, the search results right margin, and the width of the sidebar.

Replace the hardcoded values for all of these with reused variables and
math. Hopefully this will be less fragile.