Implement h group write permission function
This is our implementation of the function that checks if the currently
logged-in user is allowed to create an annotation in the given group.

This already takes the new group access flags, specifically
`writeable_by` into account. There are ideas floating around that our
"__world__" group should be a stored group as well, until then there is
a special implementation for when the specified group is "__world__"
where we check that the user has the same authority as
`request.auth_domain`.