Fix inline form submission in IE
For <form> elements that have an empty "action" attribute, IE submits
the form to the correct location when using normal submission but
returns the wrong URL when reading the form's `action` property. This
URL was used when submitting the form via fetch().

One way to fix this would be to always specify an absolute URL for the
form in the deform template. I've opted to add a workaround in JS to
avoid the code breaking if it is ever used with forms not rendered by
deform.

Fixes #3953