Change memex annotation create check for group permission
This is now using the configured groupfinder and checks with the Pyramid
ACL system if the given request has the permission to create an
annotation in the group that is returned from the groupfinder.

Since we want to support implementations that don't have the concept of
a group we will allow the annotation creation if there is no
groupfinder configured.