Remove a couple of unneeded uses of the Jinja2 "safe" filter
This is a scary filter to see in a template. In general, values should
be clearly marked as safe by the code that handles them *before* they
get to the templates.

This commit removes the use of the "safe" Jinja2 filter from a couple of
places, and adds some comments explaining when and where user input gets
escaped.