Refactor groupfinder check
All group related code is now in `memex.groups`, including an
encapsulated function for finding a group `memex.groups.find`.

We now also have a memex-default groupfinder set which returns a
`memex.groups.DefaultGroupContext` which has the ACL for a `__world__`
group where it allows writing to authenticated requests, and allows
reading for everyone. But it also has an ACL that denies all permissions
if the group id is not `__world__`. This forces all annotations to be in
the group `__world__` if there is no custom groupfinder configured.