Hardcode h letter color into logo SVG
Hardcode the white color of the "h" letter in the logo.svg file.

Previously this SVG file contained the logo's rounded rect / speech
bubble shape with the letter "h" cut out of the middle of it. That is,
the "h" itself was not a path object or any other kind of object in the
SVG, it was just formed of the negative space created by the speech
bubble shape with an "h" cut out of the middle of it.

This meant that the "h" was transparent, and the only reason that it
appeared white on the site was that the part of the page behind it was
white.

By hardcoding the white color into the "h" shape rather than leaving it
transparent, we enable the "h" to still appear white even when the logo
is displayed on non-white backgrounds.

This hardcoding is done by adding an opaque white rectangle to the SVG
file, behind the "h"-shaped hole in the speech bubble shape.