Fix blurry h logo icon in Firefox
The h logo svg icon in the top left of activity pages was sometimes
rendering blurry in Firefox.

This is because the svg image is rendered at half-pixel boundaries and,
unlike other browsers (e.g. Chrome), Firefox does not move the image
onto full pixel boundaries, but instead renders a blurry image.

See the discussion starting here:

https://github.com/hypothesis/h/pull/4215#issuecomment-267012849

Setting `transform: translateZ(0px)` on the <svg> fixes the issue.

Even though the blurriness affects all svg icons, we apply the fix only to the
h logo icon. The problem is less noticeable for other icons and we think
it may be detrimental to performance to apply the fix to a large number
of icons on the page (we haven't tested this theory).