Move annotation ACL to memex.resources.AnnotationResource
As part of introducing new kinds of group, we need to be able to have
an annotation's ACL depend on the ACL of the group it is in. This is
problematic if the ACL lives on the Annotation model because:

1. In order to retrieve the group we need to use memex.groups.find,
   which takes the request as its first argument.
2. The __acl__ method does not have access to the request.
3. There's no easy way to pass the request to the constructor of a
   SQLAlchemy mapped class.

The most straightforward way to solve this problem seems to be to
introduce a wrapper class, AnnotationResource, which acts as the request
context for views operating on annotations, and thus which holds the ACL
for the annotation.

This commit does not change any of the access control rules for
annotations, it merely introduces the AnnotationResource wrapper class
and updates the various views that previously received Annotation
objects as context to receive AnnotationResource objects instead.