Add a footer to activity pages
Like the navbar at the top of the pages, implement the footer as a Jinja2
panel that can easily be included in any page's template.

Add the footer panel to the search and user accounts pages (the same pages that
the navbar panel appears on).

Change the layout of the main `<body>` element in activity pages to a
column-direction flex box whose children are the `<header>`, the main content
`<div>`, and the new `<footer>`. This enables us to easily push the footer to
the bottom of the viewport, even when the page contents are not as tall as the
viewport, by simply putting `margin-top: auto` on the `<footer>`.

The footer CSS is designed to work well on a wide range of screen sizes
(phone, tablet, desktop...) while also not breaking if we add, edit or remove
footer items in the future, or if the user has changed the font size in their
browser.