Wrap page contents in <main> element
Wrap the main contents of every page that uses the base.html.jinja2 layout in a
`<main>` element.

This is going to enable us to use `display: flex`
(with `flex-direction: column`) on the `<body>` element. We need `<body>` to
be a flex container so that we can add a sticky, variable-height footer to some
pages, using this technique: https://philipwalton.github.io/solved-by-flexbox/

Normally making `<body>` a flex container would mess up the widths of the
content elements of some pages, because a flex container treats widths
differently to a non-flex container. (In particular, it doesn't seem to treat
`max-width` in the same way.) By wrapping the contents of every page in a
`<main>` we ensure that this `<main>`, and not the contents of the page, will
be subject to the `<body>`'s new flex layout, and avoid any unwanted changes to
the widths of page contents.

Using `<main>` also adds some semantic information.