Use the version of MutableList shipped with SQLAlchemy
When we first added support for in-place mutation of the Annotation tags
field, SQLAlchemy 1.1 hadn't yet been released, so we needed to copy
their implementation of MutableList into our code.

SQLAlchemy 1.1 has since been released, and we're already using it in
production.

This commit removes our "vendored" copy of MutableList in favour of
using the one in sqlalchemy.ext.mutable.

This ensures that in-place changes to the tags list will be detected by
the session:

     $ hypothesis --dev shell

     ...

     In [1]: ann = session.query(m.Annotation).first()

     In [2]: ann
     Out[2]: <Annotation 3WOBzPqXRH2n4wvubCPh7g>

     In [3]: ann.tags
     Out[3]: [u'3860']

     In [4]: session.dirty
     Out[4]: IdentitySet([])

     In [5]: ann.tags.append('helo')

     In [6]: session.dirty
     Out[6]: IdentitySet([<Annotation 3WOBzPqXRH2n4wvubCPh7g>])