Filter out marked-as-deleted annotations from index
We will stop deleting annotations from the index, to be able to re-index
with zero downtime. The deleted annotations will just be overridden in
the index with `{"deleted": true}`. We need to filter these out on every
single query.

The easiest way to achieve that is to check that the field doesn't
exist. In Elasticsearch 1.5 this can be done with a [missing
filter][missing-filter].
But this got removed in later versions of Elasticsearch, so to not
introduce another dependency on the old Elasticsearch version we are
using, this is done as a "not exists" filter with `bool`, `must_not`,
and `exists` (as described in the most recent, 5.1, version of the
documentation, see [ways around the removed missing
query][missing-filter-workaround]).

[missing-filter]:
https://www.elastic.co/guide/en/elasticsearch/reference/1.5/query-dsl-missing-filter.html
[missing-filter-workaround]:
https://www.elastic.co/guide/en/elasticsearch/reference/5.1/query-dsl-exists-query.html#_literal_missing_literal_query