Don't immediately delete annotations
Removing the annotation from the database synchronously (i.e. in the
DELETE request) is problematic, as it complicates the process of
streaming that delete to WebSocket users.

This commit adds a "deleted" property to the Annotation model (added to
the table in 11f8aba) and updates the AnnotationResource ACL so that
nobody has any permissions on a deleted annotation.

This ensures that the annotation cannot be returned in any direct-access
request. The preexisting search index code will ensure that the
annotation is not returned in search results.