Allow to set op_type when using BatchIndexer
When re-indexing all annotations from scratch into a new index we want
to make sure not to override any actions that have been written to the
new index already. To do this we want to set the op_type to `create`
which fails when the annotation already exists in the index.
We also use the BatchIndexer where we want to use the `index` op_type,
so we need to make sure that this is configurable.