Fix reindexing's unholy memory consumption
The reindexing process started consuming slightly ridiculous quantities
(GB) of memory, because in b0914fe we switched from loading a list of
all the updated times of annotations to loading a list of all Annotation
objects, albeit with most column values deferred.

While switching back to the prior approach would fix the problem for
now, it is still O(N) in the number of annotations, just with a smaller
constant factor.

This commit fixes the problem by switching to an altogether different
way of calculating the windows we use to stream all annotations. This
version is adapted from a recipe supplied by the SQLAlchemy maintainers:

  https://bitbucket.org/zzzeek/sqlalchemy/wiki/UsageRecipes/WindowedRangeQuery

This has the advantage of doing the window calculations on the server,
so keeps the memory in use much more strongly bounded.