Remove special-case handling of deletes in streamer
Until now it has been necessary to special-case the handling of
"delete" activity when it flows through the streamer, because by the
time the message gets to the streamer code, the annotation has been
deleted.

As of e2e6dda this is no longer true: deleted annotations remain in the
database for up to 70 minutes (a job runs every hour but has a 10-minute
buffer, so an annotation deleted 9m59s before the job runs will have to
wait until the following run), which should be plenty of time for the
streamer to process the message.