Measure coverage of test code as well
It can be very useful to see and check that the coverage of test code is
as we expect. Without this feedback, it's easier for bugs like that
fixed in https://github.com/hypothesis/h/pull/4258 to creep in.

While this does somewhat artificially inflate our coverage numbers, I
think the value of the feedback is worthwhile.