Move transaction handling from memex.eventqueue to h
This small piece of code in memex was the only part of memex that knew
anything about the request transaction. Given that it is otherwise
possible for memex to operate without knowledge of transaction handling,
and that only one piece of code actually needs a valid database session
in these after-request handlers, it seems sensible to just move the
necessary context manager to where it's needed.