Disable click-tracking for account activate link
Mandrill's click-tracking system rewrites links in emails, and the
resulting link is HTTP (not HTTPS), which means that a MitM attacker
could easily sniff the link.

While activating the user's account is not a particular dangerous thing
for an attacker to be able to do, we might in future choose to log the
user in immediately after activation, so it seems appropriately cautious
to disable click-tracking on this link.