Disable click-tracking for reset password link
Mandrill's click-tracking system rewrites links in emails, and the
resulting link is HTTP (not HTTPS), which means that a MitM attacker
could easily sniff the link.

An attacker who can sniff the reset password link can take over the
user's account, so it's important that we disable click-tracking for
this link.