Rework AuthFilter
We have moved the filtering on the groups into the new groups filter
that lives in `h`. The memex auth filter will now only check for the
`shared` field. It filters out private annotations of other users, if
the request is done on behalf of a logged-in user. Otherwise it filters
out all private annotations.