Fix renaming user to equivalent username
It was not possible to rename a user, e.g. "bob.smith", to a new
username that reduces to the same uid, e.g. "Bob.Smith", because the
rename user service would think that the username "Bob.Smith" already
exists.

Fix it to allow renaming "bob.smith" to "Bob.Smith" but *don't* allow
renaming a different user, e.g. "fred", to a username such as
"Bob.Smith" that's equivalent to the other "bob.smith" user's username.