Use the best DocumentURI for Document.web_uri
Denormalize the "best" DocumentURI.uri that we have into the
Document.web_uri field, instead of just taking the first one.

This fixes an issue that, depending on what order the DocumentURIs are
processed in, a rel-shortlink URI (for example) might be used instead of
a self-claim or canonical URI. We want to display the best available URI
for a document, a self-claim or canonical one if possible, to users.

The first time a given document is annotated consider all of the
DocumentURIs that we've generated from the document metadata sent by the
client and:

- Use a self-claim URI if there is one
- Use a rel-canonical URI if there's no self-claim
- Use any other http(s) URI if there's no self-claim or canonical
- Set web_uri to None if there are not http(s) URIs for the document