Update web_uri whenever updating document metadata
Update the value of the document.web_uri field whenever updating the
document metadata, not just the first time.

This means that, for example, if a document is annotated and then the
same document is annotated again and the second time the client sends
different document URLs then the document's web_uri will be updated as
necessary to reflect the most recently received URLs.