Add a reference to our branch naming convention
I removed the previous guidance in #4306 because it was no longer being
applied, and @robertknight pointed out that @nickstenning had proposed this new
convention on an internal email list[1]. Given we now appear to have adopted
this convention, it makes sense to update the documentation for new starters.

I've reduced the strength of this to a recommendation, because I don't think
it's reasonable to insist on a branch naming convention for external
contributors; internally, of course, we can insist as hard as we like.

[1]: https://groups.google.com/a/list.hypothes.is/forum/#4D6A209F-6444-40EA-ACF4-753B07F7AB5D