Simplify database and search index initialization
This commit removes the automatic (and hence racy) initialisation of
database tables and search index, in favour of a `hypothesis init`
command which must be called explicitly.

This simplifies the integration of memex into h (no more overriding of
memex's automatic search index initialization), and various other parts
of the application boot code, at the cost of requiring an explicit
initialisation of the database before first use.

In addition, this commit adds support for detecting if a database is
managed by Alembic, and skips automatic database initialisation in that
case. This should prevent a number of confusing behaviours which crop up
for developers when new tables are added to the database.

For development environments, this should remain largely transparent to
the user, thanks to an additional call made by the Makefile's `dev`
target.

The following settings are removed with this commit:

- MODEL_CREATE_ALL (h.db.should_create_all)
- MODEL_DROP_ALL (h.db.should_drop_all)
- SEARCH_AUTOCONFIG (h.search.autoconfig)

Fixes #4290.