Refactor annotation presenters to get links from resource
This will allow for a better API when creating a presenter, as the
resource already has the group service, so it's better to also pass the
links service to the resource instead of directly to the presenter.

This way the presenter only needs to know about the resource and nothing
else.