Make userid arg to Token mandatory
You can't create a Token without a userid (you can create the Python
object fine, but if you add it to a db session and then try to flush or
commit that session you will get an IntegrityError - userid is not
nullable in the database).

Make the userid arg a positional arg on the Token class so that its
mandatoriness is explicit and so that you can't create an invalid ORM
object.