Allow to pass access token as HTTP basic auth password
We first check for the access token in the `Authorization: Bearer
{token}` header, if we can't find it there we fall back and check if it
has been supplied as the HTTP basic auth password. We ignore the
username, but it needs to be set in some environments, especially when
the username and password is part of the url
(ie. `https://:password@host.com` doesn't seem to work in Firefox, or
curl).