Don't show `__world__` group in 3rd-party profiles
This group only makes sense when the user's authority is the same as the
request's `auth_domain`.

This also involved changing a bunch of the existing tests to make sure
the fake user was getting passed in with the correct value for the
`authority` field.

The new tests are only for the `profile` function, because we're only
going to call the `model` function for first-party requests (anonymous
or authenticated), both of which are going to show the `__world__`
group, which is already covered in other tests.