Refactor request.auth_token into a callable
A later commit in this branch will introduce fetching of an access token
from a given query string parameter.
To support fetching both the default token (from the `Authorization`
header), and from a given query string parameter, we have to refactor
this `request.auth_token` property to be a function call.

Since the calling code will be able to specify the key of the query
string parameter, and within the same request we might even want to get
different auth tokens. This means that we can't just use the pyramid
`reify` functionality for request properties anymore.

To prevent us to load the same token from the database multiple times
within a request this commit includes a new `AuthTokenFetcher` which is
a callable object that takes care of caching the fetched token
internally.