Include `public` field for all groups in profile
For the sake of simplicity in the API, it makes sense to include this
field for all groups, not just for the single group for which it's
currently true.

With the addition of groups under another authority, we're going to need
to add more logic to this, hence adding the extra tests around it to
make sure the logic doesn't break existing behaviour.