Refactor OAuthService to support multiple request types
We're going to be adding a second type of POST request to the /api/token
endpoint to go with the jwt-bearer grant_type requests that we currently
have: refresh_token grant_type requests that exchange a refresh_token for
a new access token, instead of exchanging a grant token for an access token.

So OAuthService will need to be able to verify both jwt-bearer and
refresh_token requests.

This commit refactors OAuthService a little to make way for adding in the
refresh_token request support.

Instead of OAuthService.verify_jwt_bearer(), user code will now just call
OAuthService.verify_token_request(), which first checks the grant_type of
the request (jwt-bearer or refresh_token) and then calls either
_verify_jwt_bearer() or _verify_refresh_token() (not yet implemented).