Make the User factory into a ModelFactory (#4335)
This factory was originally created as a subclass of the base `Factory`
(see commit a928eb5e) rather than our custom `ModelFactory`, which
meant it would not add the objects it generated to the current session.
This is different behaviour to the other factories, which can be a
little confusing. We also have the `build` method, which achieves the
same outcome.

I've gone through the tests that invoke this factory, and I think I've
got them all working as intended. Where the tests rely on patching out
method calls, I've assumed that this means the behaviour of not adding
the new objects to the session is intended, and changed the call to a
`build` call. Where the factory is immediately followed by `session.add`
calls, I've removed them as they now happen automatically. In other
cases, I've mostly assumed that adding the new object to the session or
not has no effect on the test, and have left the method as it stands.