Generate 'links' object in /api route output from API view config metadata
Generate the 'links' object returned by the /api route dynamically from
metadata specified with the `api_config` decorator instead of
hard-coding the list of routes in the `api.index` function.

Metadata for all views configured using the `api_config` decorator which
specify a `link_name` kwarg are registered in the `api_links` array
property on the application's registry.

The `/api` route then enumerates these links in order to generate the
response.

This allows consumers of the memex package, such as h, to extend the set
of routes by defining additional views using the `api_config` decorator
and specifying a dotted path in the `link_name` param.