Refactor auth token parsing and loading
Previously, both getting the auth token from the request headers, and
then either loading it from the database or trying to decode a legacy
JWT token happened in the same place.

We now want to support having the client pass the auth token in the
`access_token` request query parameter as well, but only for the
websocket server. To achieve this we need to split up the parsing and
validating of the token.