Remove `aud` from legacy client JWT token
The audience check was a protection to ensure that we don't use a token
destined for someone else, but we only ever generate these JWT tokens
for ourselves. So there is no need for checking the audience anymore.
This should not compromise security, because as long as the server-side
client secret is not compromised, there is no way to tamper with the JWT
token.

This will allow us to use the legacy JWT tokens for authenticating
websocket connections when the web service and the websocket service run
on different hosts (in localhost we have: `http://localhost:5000` for
the web service, and `http://localhost:5001` for the websocket service).