Move calculation of Token TTLs into Token class
Add a Token.ttl @property that returns the token's time to live (from
now).

This removes the dependency on h.services.oauth.TOKEN_TTL from
h.views.api_auth.

This also means that the expires_in times that the API returns are now
correct. Previously if the API had returned a token with a one hour life
time that had been created 30 mins ago, the expires_in that the API
returns would have indicated one hour instead of 30 mins. Now it will
indicate 30 mins. (In practice the API currently only ever returns newly
created tokens and not old ones, but if it changes to sometimes return
old ones in future then the expires_in will now be correct.)

In practice, with the current behaviour that we always create a new
token in the db immediately before returning an API response, this
change means that expires_in will now be 3599 instead of 3600 because it
rounds down to get the number of seconds as an int.