Support unauthenticated `profile` requests
This functionality is necessary so that we can display a list of public
groups and use a default set of feature flags before a user has logged
in. This needs to work both for the default authority and for a
third party.

There was some discussion around whether to expect the `authority`
parameter, even for authenticated requests, and fail if the value did
not match the user's authority. The main argument behind this was to
provide separation between the different URLs used for profiles from
different authorities. However, in the interests of unblocking the
client component of this work, and given a piece of suggested upcoming
investigation into providing entirely separate API namespaces for
third parties, we've gone with the simplest option for the moment.