Add fake request class to remove test duplication (#4351)
A lot of these tests were building up the same set of nested mock
objects as one another, with different sets of groups and features.  For
the sake of simplicity (and at @nickstenning's suggestion), this sets up
a `FakeRequest` class that sets up the request methods that the `model`
and `profile` functions expect, and that knows how to customise its own
internals in a few set ways. This means the user fixtures are no longer
referenced and can be removed.

I found that this approach didn't work very nicely with the
heavily-parametrized tests around user preferences, so I split these
into separate tests for authenticated and unauthenticated requests,
which removed a good chunk of the complexity.