Display publisher groups in the profile API
This fixes the misunderstanding whereby a profile for a third-party
account would not display the publisher group for the user's authority,
because a publisher group is unjoinable rather than being automatically
joined by eligible users.

This also frees up the `session` module from having to know the
difference between an authority that uses a publisher group and the
authority that uses the `WorldGroup` class, since this is now handled by
the new `AuthorityGroupService`.