Run collectd inside docker container for system metrics
In order to get system metrics from h running on Skyliner, we need to
run collectd inside the container.

This adds the exact same collectd configuration as we currently use on
the old infrastructure instances, except for the removal of the nptd
plugin which doesn't work due to [a
bug](https://github.com/collectd/collectd/issues/932).

The `start-collectd` script will configure collectd to send the metrics
to graphite if the `GRAPHITE_HOST` environment variable is set,
otherwise it will configure the basic `write_log` plugin.
Similar to that it tries to get the EC2 instance id through the metadata
API, if it isn't available it will fallback to the local hostname
(calling `hostname`).